Moves file ops method to dedicated file #14
