+++ cmd_createitem.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
 	"log"
-	"os"
 
 	common "github.com/apiheat/akamai-cli-common"
 	service "github.com/apiheat/go-edgegrid/v6/service/netlistv2"
@@ -60,27 +58,4 @@ func addItemsToNetlist(c *cli.Context) error {
 
 	return nil
 
-}
-
-// readLinesFromFile reads a whole file into memory
-// and returns a slice of its lines.
-func readLinesFromFile(path string) ([]string, error) {
-
-	file, err := os.OpenFile(path, os.O_RDONLY, os.ModeAppend)
-	if err != nil {
-		if os.IsNotExist(err) {
-			fmt.Print("File Does Not Exist: ")
-		}
-		fmt.Println(err)
-		return nil, err
-	}
-
-	defer file.Close()
-
-	var lines []string
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		lines = append(lines, scanner.Text())
-	}
-	return lines, scanner.Err()
 }
